Adds error message association to field set, moves CorrespondenceFieldset to logical namespace, cleans up files some associated files
