+++ service-front/module/Application/tests/Form/Lpa/CorrespondenceFieldsetTest.php
@@ -2,22 +2,22 @@
 
 namespace ApplicationTest\Form\Lpa;
 
-use Application\Form\Lpa\CorrespondenceFieldset;
+use Application\Form\Fieldset\Correspondence;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 
 class CorrespondenceFieldsetTest extends MockeryTestCase
 {
     public function testNameAndInstance()
     {
-        $fieldSet = new CorrespondenceFieldset();
+        $fieldSet = new Correspondence();
 
-        $this->assertInstanceOf('Application\Form\Lpa\CorrespondenceFieldset', $fieldSet);
+        $this->assertInstanceOf('Application\Form\Fieldset\Correspondence', $fieldSet);
         $this->assertInstanceOf('Laminas\Form\Fieldset', $fieldSet);
     }
 
     public function testElements()
     {
-        $fieldSet = new CorrespondenceFieldset();
+        $fieldSet = new Correspondence();
 
         $this->assertInstanceOf('Laminas\Form\Element\Checkbox', $fieldSet->get('contactByEmail'));
         $this->assertInstanceOf('Laminas\Form\Element\Checkbox', $fieldSet->get('contactByPhone'));
@@ -32,7 +32,7 @@ class CorrespondenceFieldsetTest extends MockeryTestCase
             'some-field' => 'A big error message',
         ];
 
-        $fieldSet = new CorrespondenceFieldset();
+        $fieldSet = new Correspondence();
 
         $fieldSet->setMessages($testMessages);
